Project layout restructure
Changed file names, simplified directory structure.
