+++ driven/driven.py
@@ -4,7 +4,7 @@ import click
 @click.command()
 @click.option('--as-cowboy', '-c', is_flag=True, help='Greet as a cowboy.')
 @click.argument('name', default='world', required=False)
-def main(name, as_cowboy):
+def cli(name, as_cowboy):
     """Driven is a full-fledged bioinformatics suite geared towards easy integration of multiple types of big data to make reasonable and interesting biological conclusions."""
     greet = 'Howdy' if as_cowboy else 'Hello'
     click.echo('{0}, {1}.'.format(greet, name))
